LayoutTests/http/wpt/service-workers/service-worker-crashing-while-fetching.https-expected.txt

CONSOLE MESSAGE: Service Worker context closed
CONSOLE MESSAGE: Cannot load .


PASS Setup worker
PASS Service worker process crashes while handling a fetch event
